changing the start method with the @Container annotation to @BeforAll, moving the test container to the integration folder, moved everything related to the test container into a separate class and inherited it where it is needed, added Loger
