+++ build.gradle
@@ -11,7 +11,11 @@ version '1.0-SNAPSHOT'
 repositories {
     mavenLocal()
     mavenCentral()
+
 }
+
+
+
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter:3.0.6'
     implementation 'org.springframework.boot:spring-boot-starter-web:3.0.6'
@@ -19,6 +23,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation:3.0.6'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa:3.0.6'
     implementation 'org.springframework.boot:spring-boot-starter-cache:3.0.6'
+    implementation 'org.springframework.boot:spring-boot-starter-aop:3.0.6'
     implementation 'org.hibernate:hibernate-entitymanager:6.0.0.Alpha7@pom'
 
     implementation 'org.modelmapper:modelmapper:2.4.4'
@@ -29,6 +34,8 @@ dependencies {
     implementation 'org.postgresql:postgresql:42.5.1'
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.14.2'
     runtimeOnly 'org.aspectj:aspectjweaver:1.9.9.1'
+
+    implementation files('../lib/loger.jar')
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test:3.0.6'
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.9.1'
