--- src/main/java/ru/clevertec/ecl/aop/loger/ExcludeLog.java
@@ -1,12 +1,0 @@
-package ru.clevertec.ecl.aop.loger;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface ExcludeLog {
-}
