+++ src/test/java/ru/clevertec/ecl/common/integration/UserPostgresQLRepositoryTest.java
@@ -1,4 +1,4 @@
-package ru.clevertec.ecl.common.repository;
+package ru.clevertec.ecl.common.integration;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -7,13 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
 import org.springframework.test.context.jdbc.Sql;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 import ru.clevertec.ecl.repository.user.UserRepository;
 
 
@@ -21,32 +16,13 @@ import ru.clevertec.ecl.repository.user.UserRepository;
 @Testcontainers
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @DisplayName("User Repository Test")
-public class UserPostgresQLRepositoryTest {
+public class UserPostgresQLRepositoryTest extends TestContainerInitializer{
 
     @Autowired
     private  UserRepository userRepository;
 
     @Autowired
     private  TestEntityManager testEntityManager;
-
-
-
-    private static final DockerImageName PostgresQL_IMAGE = DockerImageName
-            .parse("postgres:14.3");
-    @Container
-    private  static final PostgreSQLContainer<?> postgresSQLContainer =
-            new PostgreSQLContainer<>(PostgresQL_IMAGE)
-            .withDatabaseName("test")
-            .withUsername("root")
-            .withPassword("root")
-            .withExposedPorts(5432);
-
-    @DynamicPropertySource
-    private static void setProperties(DynamicPropertyRegistry registry){
-        registry.add("spring.datasource.url", postgresSQLContainer::getJdbcUrl);
-        registry.add("spring.datasource.password", postgresSQLContainer::getPassword);
-        registry.add("spring.datasource.username", postgresSQLContainer::getUsername);
-    }
 
     @Test
     @Sql("/INITIAL_ALL_DB_SCRIPT.sql")
@@ -56,6 +32,4 @@ public class UserPostgresQLRepositoryTest {
         testEntityManager.getEntityManager().getTransaction().commit();
        Assertions.assertEquals(1, activeUserName);
     }
-
-
 }
