+++ src/test/java/ru/clevertec/ecl/common/service/UserServiceImplTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import ru.clevertec.ecl.common.mapper.serviceMapper.UserServiceImplTestMapper;
+import ru.clevertec.ecl.common.utill.serviceMapper.UserServiceImplTestMapper;
 import ru.clevertec.ecl.dto.user.UserDtoRequest;
 import ru.clevertec.ecl.entity.user.User;
 import ru.clevertec.ecl.repository.user.UserRepository;
